Fix `update` behavior
Our previous logic was creating two separate sessions when trying to
update an object. Because we wrap our object modifications with a
separate `with_transaction` the wrong session was picking up our object
changes. As a result, at merge time the behavior was a no-op.

Update the API contract of the merge function to pass in the session.
This will allow client callers to easily override the merge behavior
(much like in the original microcosm-postgres logic) while respecting
the session that the base store is opening within `update` and
`update_with_diff`.