Use fastapi default response (#34)
The latest version of microcosm-fastapi & fastapi sees a `ValueError: invalid literal for int() with base 10: 'default'` crash on service initialization if we use the docs factory to provide automatic documentation.

This was caused by the default response decorator, which includes this defaults key:
https://github.com/globality-corp/microcosm-fastapi/blob/master/microcosm_fastapi/factories/fastapi.py#L88

Fastapi broke this logic by introducing an additional assert in tiangolo/fastapi@c431202 that ensures response codes are numeric. The "default" string is being treated as the response code here, which cases the error.

Here we change the default response code to be 200, along with the payload that fastapi includes in their schemas by default.